update avatar onChange
